Work around weird zcat(1) behaviour on Mac OS
The system /usr/bin/zcat on Mac OS replicates an oddity of traditional
pre-gzip Unix compress(1): it assumes it should append ".Z" to the name of
the file you give it (which means that it then fails when that file doesn't
exist). So just use "gzip -dc" directly on all platforms.